+++ apps/web/src/api.ts
@@ -4,6 +4,10 @@ import type { EdenApp } from "server";
 // In production, API requests will be proxied through Caddy at /api/*
 // In development, they go directly to localhost:3000
 const baseUrl =
-	process.env.NODE_ENV === "production" ? "localhost/api" : "localhost:3000";
+	process.env.NODE_ENV === "production"
+		? `http://${window.location.host}/api`
+		: "localhost:3000";
+
+console.log(window.location.host);
 
 export const api = treaty<EdenApp>(baseUrl);