+++ CHANGELOG.md
@@ -1,4 +1,23 @@
+## 0.22.7 - 21 Oct 2025
+
+## 0.22.6 - 21 Oct 2025
+
+## 0.22.5 - 21 Oct 2025
+
+## 0.22.4 - 21 Oct 2025
+
+Bugfix:
+- ports for production!
+
+## 0.22.3 - 20 Oct 2025
+
+Bugfix:
+- redis.conf
+
 ## 0.22.2 - 20 Oct 2025
+
+Bugfix:
+- Include redis-stack.conf
 
 ## 0.22.1 - 17 Oct 2025